Simplify some Cairo drawing
Instead of generating paths twice, once for filling and once for
stroking, use the path-preserving `fill_preserve` operator instead. As
well, do the filling before stroking for consistency, or else the stroke
might be accidentally covered, as occurred with the child rings.

Also, simplify the map marker which seems to have drawn the elliptical
portion twice.